chore: rename editor.ts to EditorUtils.ts
