+++ packages/plugin-core/src/commands/CopyNoteRef.ts
@@ -16,7 +16,7 @@ import { PickerUtilsV2 } from "../components/lookup/utils";
 import { DENDRON_COMMANDS } from "../constants";
 import { ExtensionProvider } from "../ExtensionProvider";
 import { clipboard } from "../utils";
-import { EditorUtils } from "../utils/editor";
+import { EditorUtils } from "../utils/EditorUtils";
 import { VSCodeUtils } from "../vsCodeUtils";
 import { getEngine } from "../workspace";
 import { BasicCommand } from "./base";
