+++ packages/plugin-core/src/utils/index.ts
@@ -1,2 +1,2 @@
-export * from "./editor";
+export * from "./EditorUtils";
 export * from "./StartupPrompts";
